% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{determine_significant_features}
\alias{determine_significant_features}
\title{determine_significant_features}
\usage{
determine_significant_features(X, Y, orig_model, num_permutations, talpha,
  inside_num_permutations, inside_talpha)
}
\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix. Must be numeric}

\item{orig_model}{- original model.}

\item{num_permutations}{- Number of permutation iterations.}

\item{talpha}{- test alpha to use as cutoff}

\item{inside_num_permutations}{- the smaller number of permutations used to efficiently rule out insignificant features. Should be less than num_permutations.}

\item{inside_talpha}{- corresponding test alpha for ruling out loadings that are not close to being significant. This is purely for efficient calculations.}
}
\value{
List containing \item{sig_inxs}{Significant indices} \item{not_sig_inxs}{Indices that are not significant} \item{significant}{0/1 bit vector for significant variables} \item{p_permuted}{the permuted loading values}
}
\description{
Permute the values of each variable and recompute its loadings. Compare this distribution of loadings with the original loading for each variable.
}
\examples{
X <- rand(10,10)
Y <- rand(10,1)
model <- opls(X,Y,1)
res <- determine_significant_features(X,Y,model,500,0.05,100,0.2)
}

