% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{opls}
\alias{opls}
\alias{opls-package}
\title{opls.}
\usage{
opls(X, Y, num_OPLS_fact)
}
\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix. Must be numeric}

\item{num_OPLS_fact}{- Integer specifying the number of OPLS orthogonal components.}
}
\value{
the found opls model
}
\description{
opls.

This allows you to create an OPLS model if you know the number of orthogonal components.
}
\examples{
model <- opls(rand(10,10),rand(10,1),1)
}

