% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{opls_CV}
\alias{opls_CV}
\title{opls_CV}
\usage{
opls_CV(X, Y, num_OPLS_fact, folds)
}
\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix. Must be numeric}

\item{num_OPLS_fact}{- Integer specifying the number of OPLS orthogonal components.}

\item{folds}{- Number of k-fold cross-validation groups or -1 for leave one out cross-validation.}
}
\value{
List containing \item{Q^2}{Cross-validated R^2} \item{Q2s}{One for each iteration} \item{press}{residual calculation used in Q^2 calculation} \item{accuracy}{standard accuracy that is only relevant if this is a classification problem}
}
\description{
This allows you to create an OPLS model if you know the number of orthogonal components.
}
\examples{
res <- opls_CV(rand(10,10),rand(10,1),1,-1)
}

