% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{permutation_test}
\alias{permutation_test}
\title{permutation_test}
\usage{
permutation_test(X, Y, num_OPLS_fact, num_permutations, folds)
}
\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix. Must be numeric}

\item{num_OPLS_fact}{- number of orthogonal OPLS factors}

\item{num_permutations}{- number of permutation for the randomization test.}

\item{folds}{- number of folds for CV}
}
\description{
Internal helper function. Do not use.
}

