% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{run_det_sig}
\alias{run_det_sig}
\title{run_det_sig}
\usage{
run_det_sig(X, Y, orig_model, N, variables)
}
\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix. Must be numeric}

\item{orig_model}{- original model.}

\item{N}{- Number of permutation iterations.}

\item{variables}{- list of variables to test}
}
\description{
This is an internal helper function. Do not call directly.
}

