% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{run_opls}
\alias{run_opls}
\title{run_opls}
\usage{
run_opls(X, Y, num_permutations, CV, min_num_OPLS_fact = 0)
}
\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix. Must be numeric}

\item{num_permutations}{- Number of permutation iterations.}

\item{CV}{- Number of k-fold cross-validation groups or -1 for leave one out cross-validation.}

\item{min_num_OPLS_fact}{- minimum number of OPLS orthogonal components. Default 0.}
}
\value{
List containing \item{model}{Final OPLS model} \item{Q2}{Cross-validated R^2} \item{accuracy}{standard accuracy that is only relevant if this is a classification problem} \item{num_OPLS_fact}{number of orthogonal components} \item{permutation_Q2s}{Q2 from each permutation test} \item{pvalue}{p-value using the permutation test}
}
\description{
This will perform OPLS analysis including cross-validation and a permutation test.
}
\examples{
res <- run_opls(rand(10,10),rand(10,1),100,-1)
}

